+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next Release
+
+* Corrected issues with duplicate JavaScript being rendered for each block.
+* Corrected missing media from `StructBlockField` and `StreamBlockField`
+* Improved StreamValue to convert more items to UnknownBlock if they cannot be parsed
+* Improved StreamValue to throw away some data to try to retain the block if there is a validation error of the block value.
+
 ## 0.0.7
 
 * Allowed `StreamValue` to substitute a block whose `to_python()` function raises a `ValidationError` with an `UnknownBlock`.